removed Old Enclave Client and Azure Enclave Client
