+++ packages/syft/src/syft/client/enclave_client.py
@@ -1,13 +1,34 @@
+# future
+from __future__ import annotations
+
 # stdlib
 from typing import Optional
+from typing import TYPE_CHECKING
 
 # third party
 from typing_extensions import Self
 
 # relative
+from ..client.api import APIRegistry
 from ..serde.serializable import serializable
+from ..types.syft_object import SYFT_OBJECT_VERSION_1
+from ..types.syft_object import SyftObject
+from ..types.uid import UID
 from .api import APIModule
 from .client import SyftClient
+from .connection import NodeConnection
+
+if TYPE_CHECKING:
+    # relative
+    from ..service.code.user_code import SubmitUserCode
+
+
+@serializable()
+class EnclaveMetadata(SyftObject):
+    __canonical_name__ = "EnclaveMetadata"
+    __version__ = SYFT_OBJECT_VERSION_1
+
+    connection: NodeConnection
 
 
 @serializable()
@@ -27,3 +48,36 @@ class EnclaveClient(SyftClient):
 
     def apply_to_gateway(self, client: Self) -> None:
         return self.exchange_route(client)
+
+    def get_enclave_metadata(self) -> EnclaveMetadata:
+        return EnclaveMetadata(connection=self.connection)
+
+    def request_code_execution(self, code: SubmitUserCode):
+        # relative
+        from ..service.code.user_code_service import SubmitUserCode
+
+        if not isinstance(code, SubmitUserCode):
+            raise Exception(
+                f"The input code should be of type: {SubmitUserCode} got:{type(code)}"
+            )
+
+        enclave_metadata = self.get_enclave_metadata()
+
+        code_id = UID()
+        code.id = code_id
+        code.enclave_metadata = enclave_metadata
+
+        apis = []
+        for k, v in code.input_policy_init_kwargs.items():
+            api = APIRegistry.api_for(k.node_id, k.verify_key)
+            if api is None:
+                raise ValueError(f"could not find client for input {v}")
+            else:
+                apis += [api]
+
+        for api in apis:
+            api.services.code.request_code_execution(code=code)
+
+        res = self.api.services.code.request_code_execution(code=code)
+
+        return res
