--- packages/syft/src/syft/enclave/__init__.py
@@ -1,5 +1,0 @@
-# relative
-from .enclave_client import AzureEnclaveClient
-from .metadata import EnclaveMetadata
-
-__all__ = ["AzureEnclaveClient", "EnclaveMetadata"]
