--- packages/syft/src/syft/enclave/metadata.py
@@ -1,8 +1,0 @@
-# relative
-from ..types.base import SyftBaseModel
-
-
-class EnclaveMetadata(SyftBaseModel):
-    """Contains metadata to connect to a specific Enclave"""
-
-    pass
